Fixing crash in watchOS app
Renaming StandardCache to DiskCache. Fixing a bug where we were looking for the absoluteString of a cached file in DiskCache, not the path. Now caching artwork by album title, not playcut id, so as to increase our chances from 0 that we'll find a cache hit if a track from the same album is played later. Increasing the eviction policy of artwork to thirty days.